+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change log
 
+## v1.3.1 - Apr. 14, 2020
+
+**New features:**
+- Field aliases are not necessary while using field modifiers
+
+**Bug fixes:**
+- Where clause bitwise operations
+
+---
+
 ## v1.3 - Apr. 12, 2020
 
 **New features:**